+++ mobile/android/android-components/components/feature/prompts/src/main/java/mozilla/components/feature/prompts/PromptFeature.kt
@@ -42,6 +42,7 @@ import mozilla.components.concept.engine.prompt.PromptRequest.TextPrompt
 import mozilla.components.concept.engine.prompt.PromptRequest.TimeSelection
 import mozilla.components.concept.storage.Login
 import mozilla.components.concept.storage.LoginValidationDelegate
+import mozilla.components.feature.prompts.concept.SelectablePromptView
 import mozilla.components.feature.prompts.dialog.AlertDialogFragment
 import mozilla.components.feature.prompts.dialog.AuthenticationDialogFragment
 import mozilla.components.feature.prompts.dialog.ChoiceDialogFragment
@@ -60,7 +61,6 @@ import mozilla.components.feature.prompts.dialog.TimePickerDialogFragment
 import mozilla.components.feature.prompts.file.FilePicker
 import mozilla.components.feature.prompts.login.LoginExceptions
 import mozilla.components.feature.prompts.login.LoginPicker
-import mozilla.components.feature.prompts.login.LoginPickerView
 import mozilla.components.feature.prompts.share.DefaultShareDelegate
 import mozilla.components.feature.prompts.share.ShareDelegate
 import mozilla.components.lib.state.ext.flowScoped
@@ -107,7 +107,8 @@ internal const val FRAGMENT_TAG = "mozac_feature_prompt_dialog"
  * 'save login'prompts will not be shown.
  * @property loginExceptionStorage An implementation of [LoginExceptions] that saves and checks origins
  * the user does not want to see a save login dialog for.
- * @property loginPickerView The [LoginPickerView] used for [LoginPicker] to display select login options.
+ * @property loginPickerView The [SelectablePromptView] used for [LoginPicker] to display a
+ * selectable prompt list of login options.
  * @property onManageLogins A callback invoked when a user selects "manage logins" from the
  * select login prompt.
  * @property onNeedToRequestPermissions A callback invoked when permissions
@@ -124,7 +125,7 @@ class PromptFeature private constructor(
     override val loginValidationDelegate: LoginValidationDelegate? = null,
     private val isSaveLoginEnabled: () -> Boolean = { false },
     override val loginExceptionStorage: LoginExceptions? = null,
-    private val loginPickerView: LoginPickerView? = null,
+    private val loginPickerView: SelectablePromptView<Login>? = null,
     private val onManageLogins: () -> Unit = {},
     onNeedToRequestPermissions: OnNeedToRequestPermissions
 ) : LifecycleAwareFeature, PermissionsFeature, Prompter, ActivityResultHandler,
@@ -151,7 +152,7 @@ class PromptFeature private constructor(
         loginValidationDelegate: LoginValidationDelegate? = null,
         isSaveLoginEnabled: () -> Boolean = { false },
         loginExceptionStorage: LoginExceptions? = null,
-        loginPickerView: LoginPickerView? = null,
+        loginPickerView: SelectablePromptView<Login>? = null,
         onManageLogins: () -> Unit = {},
         onNeedToRequestPermissions: OnNeedToRequestPermissions
     ) : this(
@@ -177,7 +178,7 @@ class PromptFeature private constructor(
         loginValidationDelegate: LoginValidationDelegate? = null,
         isSaveLoginEnabled: () -> Boolean = { false },
         loginExceptionStorage: LoginExceptions? = null,
-        loginPickerView: LoginPickerView? = null,
+        loginPickerView: SelectablePromptView<Login>? = null,
         onManageLogins: () -> Unit = {},
         onNeedToRequestPermissions: OnNeedToRequestPermissions
     ) : this(
@@ -201,7 +202,7 @@ class PromptFeature private constructor(
         store: BrowserStore,
         customTabId: String? = null,
         fragmentManager: FragmentManager,
-        loginPickerView: LoginPickerView? = null,
+        loginPickerView: SelectablePromptView<Login>? = null,
         onManageLogins: () -> Unit = {},
         onNeedToRequestPermissions: OnNeedToRequestPermissions
     ) : this(