+++ mobile/android/android-components/components/feature/prompts/src/main/java/mozilla/components/feature/prompts/login/LoginPicker.kt
@@ -7,44 +7,45 @@ package mozilla.components.feature.prompts.login
 import mozilla.components.browser.state.store.BrowserStore
 import mozilla.components.concept.engine.prompt.PromptRequest
 import mozilla.components.concept.storage.Login
+import mozilla.components.feature.prompts.concept.SelectablePromptView
 import mozilla.components.feature.prompts.consumePromptFrom
 import mozilla.components.support.base.log.logger.Logger
 
 /**
- * The [LoginPicker] displays a list of possible logins in a [LoginPickerView] for a site after
+ * The [LoginPicker] displays a list of possible logins in a [SelectablePromptView] for a site after
  * receiving a [PromptRequest.SelectLoginPrompt] when a user clicks into a login field and we have
  * matching logins. It allows the user to select which one of these logins they would like to fill,
  * or select an option to manage their logins.
  *
  * @property store The [BrowserStore] this feature should subscribe to.
- * @property loginSelectBar The [LoginPickerView] view into which the select login "prompt" will be inflated.
+ * @property loginSelectBar The [SelectablePromptView] view into which the select login "prompt" will be inflated.
  * @property manageLoginsCallback A callback invoked when a user selects "manage logins" from the
  * select login prompt.
  * @property sessionId This is the id of the session which requested the prompt.
  */
 internal class LoginPicker(
     private val store: BrowserStore,
-    private val loginSelectBar: LoginPickerView,
+    private val loginSelectBar: SelectablePromptView<Login>,
     private val manageLoginsCallback: () -> Unit = {},
     private var sessionId: String? = null
-) : LoginPickerView.Listener {
+) : SelectablePromptView.Listener<Login> {
 
     init {
         loginSelectBar.listener = this
     }
 
     internal fun handleSelectLoginRequest(request: PromptRequest.SelectLoginPrompt) {
-        loginSelectBar.showPicker(request.logins)
+        loginSelectBar.showPrompt(request.logins)
     }
 
-    override fun onLoginSelected(login: Login) {
+    override fun onOptionSelect(option: Login) {
         store.consumePromptFrom(sessionId) {
-            if (it is PromptRequest.SelectLoginPrompt) it.onConfirm(login)
+            if (it is PromptRequest.SelectLoginPrompt) it.onConfirm(option)
         }
-        loginSelectBar.hidePicker()
+        loginSelectBar.hidePrompt()
     }
 
-    override fun onManageLogins() {
+    override fun onManageOptions() {
         manageLoginsCallback.invoke()
         dismissCurrentLoginSelect()
     }
@@ -58,6 +59,6 @@ internal class LoginPicker(
         } catch (e: RuntimeException) {
             Logger.error("Can't dismiss this login select prompt", e)
         }
-        loginSelectBar.hidePicker()
+        loginSelectBar.hidePrompt()
     }
 }