+++ mobile/android/android-components/components/feature/prompts/src/main/java/mozilla/components/feature/prompts/login/LoginSelectBar.kt
@@ -20,16 +20,17 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import mozilla.components.concept.storage.Login
 import mozilla.components.feature.prompts.R
+import mozilla.components.feature.prompts.concept.SelectablePromptView
 import mozilla.components.support.ktx.android.view.hideKeyboard
 
 /**
- * A customizable multiple login selection bar implementing [LoginPickerView].
+ * A customizable multiple login selection bar implementing [SelectablePromptView].
  */
 class LoginSelectBar @JvmOverloads constructor(
     context: Context,
     attrs: AttributeSet? = null,
     defStyleAttr: Int = 0
-) : ConstraintLayout(context, attrs, defStyleAttr), LoginPickerView {
+) : ConstraintLayout(context, attrs, defStyleAttr), SelectablePromptView<Login> {
 
     var headerTextStyle: Int? = null
 
@@ -48,16 +49,16 @@ class LoginSelectBar @JvmOverloads constructor(
         }
     }
 
-    override var listener: LoginPickerView.Listener? = null
+    override var listener: SelectablePromptView.Listener<Login>? = null
 
-    override fun showPicker(list: List<Login>) {
+    override fun showPrompt(options: List<Login>) {
         tryInflate().also {
-            listAdapter.submitList(list)
+            listAdapter.submitList(options)
             loginPickerView?.isVisible = true
         }
     }
 
-    override fun hidePicker() {
+    override fun hidePrompt() {
         this.isVisible = false
         loginsList?.isVisible = false
         listAdapter.submitList(mutableListOf())
@@ -76,7 +77,7 @@ class LoginSelectBar @JvmOverloads constructor(
     private var expandArrowHead: AppCompatImageView? = null
 
     private var listAdapter = BasicLoginAdapter {
-        listener?.onLoginSelected(it)
+        listener?.onOptionSelect(it)
     }
 
     override fun tryInflate(): Boolean {
@@ -93,7 +94,7 @@ class LoginSelectBar @JvmOverloads constructor(
     private fun bindViews() {
         manageLoginsButton = findViewById<AppCompatTextView>(R.id.manage_logins).apply {
             setOnClickListener {
-                listener?.onManageLogins()
+                listener?.onManageOptions()
             }
         }
         loginsList = findViewById(R.id.logins_list)